+++ Src/CSharpier.Cli/CommandLineFormatter.cs
@@ -189,7 +189,7 @@ internal static class CommandLineFormatter
 
             if (!Path.IsPathRooted(originalDirectoryOrFile))
             {
-                if (!originalDirectoryOrFile.StartsWith("."))
+                if (!originalDirectoryOrFile.StartsWith('.'))
                 {
                     originalDirectoryOrFile = "./" + originalDirectoryOrFile;
                 }
@@ -422,13 +422,13 @@ internal static class CommandLineFormatter
             return;
         }
 
-        if (!codeFormattingResult.WarningMessage.IsBlank())
+        if (!string.IsNullOrEmpty(codeFormattingResult.WarningMessage))
         {
             fileIssueLogger.WriteWarning(codeFormattingResult.WarningMessage);
             return;
         }
 
-        if (!codeFormattingResult.FailureMessage.IsBlank())
+        if (!string.IsNullOrEmpty(codeFormattingResult.FailureMessage))
         {
             fileIssueLogger.WriteError(codeFormattingResult.FailureMessage);
             return;