+++ Src/CSharpier/CSharpFormatter.cs
@@ -7,6 +7,7 @@ namespace CSharpier;
 
 internal static class CSharpFormatter
 {
+    private static readonly JsonSerializerOptions IndentedJson = new() { WriteIndented = true };
     internal static readonly LanguageVersion LanguageVersion = LanguageVersion.Preview;
 
     internal static Task<CodeFormatterResult> FormatAsync(
@@ -85,7 +86,7 @@ internal static class CSharpFormatter
                 .GetDiagnostics(cancellationToken)
                 .Where(o => o.Severity == DiagnosticSeverity.Error && o.Id != "CS1029")
                 .ToList();
-            if (diagnostics.Any())
+            if (diagnostics.Count != 0)
             {
                 compilationResult = new CodeFormatterResult
                 {
@@ -188,17 +189,14 @@ internal static class CSharpFormatter
             // SyntaxNodeJsonWriter doesn't write things indented, so this cleans it up for us
             return JsonSerializer.Serialize(
                 JsonSerializer.Deserialize<object>(stringBuilder.ToString()),
-                new JsonSerializerOptions { WriteIndented = false }
+                IndentedJson
             );
         }
         // in some cases with new unsupported c# language features
         // SyntaxNodeJsonWriter will not produce valid json
         catch (JsonException ex)
         {
-            return JsonSerializer.Serialize(
-                new { exception = ex.ToString() },
-                new JsonSerializerOptions { WriteIndented = false }
-            );
+            return JsonSerializer.Serialize(new { exception = ex.ToString() }, IndentedJson);
         }
     }
 }